Apply principle of least privilege for granting the operator permissions
ClusterRoleBinding grants the permissions on all namespaces within the
cluster, rather than on a specific namespace. After this commit, we'll
only use that when watching all namespaces.

RoleBinding grants the permissions on a specific namespace, we'll
use that when a specific set of namespaces is being watched, so
that the cluster does not get any permissions on namespaces it
is not watching, following the principle of least privilege.

Note that the ClusterRole used to reference in its subject a
ServiceAccount in the namespace being watched, this is incorrect,
as service account only exists in the namespace in which the
operator is created. This wasn't problematic in practice, as the
one ClusterRoleBinding for the namespace in which the operator
is created was already granting the permissions on all namespaces,
so there was no problem in the others not granting the permission
to the right service account. In this commit, the subject now
references the right service account in the namespace in which
the service account is created.