Follow principle of least privilege for operator permissions
ClusterRoleBinding grants the permissions cluster wide. Instead,
now we limit the permissions to just the namespaces being watched
by using a RoleBinding instead following the principle of least
privilege.

Also, the service account is created in the default namespace,
not in each watched namespace separately.
This commit also changes the service account namespace to be
the release namespace (provided with --namespace on the helm
install command), instead of always the default namespace.